fix: resolve remaining react-doctor findings in code
- Replace internal barrel imports in LiveChart/LiveChartSeries with direct
  module imports (no-barrel-import).
- Use content-derived keys for reference lines instead of the array index
  (no-array-index-key / no-array-index-as-key).
- Extract the test-only `seriesMetaSig` worklet helper to its own module so the
  component file only exports components (only-export-components).
- Convert the demo home menu from ScrollView+map to FlatList with a module-scope
  renderItem (rn-no-scrollview-mapped-list + no inline renderItem).
- Lazily init the sim bonding-curve ref (rerender-lazy-ref-init).
- Hoist the LiveChart tooltipBody out of the inline JSX prop position.

Narrow doctor.config to the genuinely-irreducible false positives, each scoped
to its specific file with rationale: fixed-size render-slot-pool keys, the
LiveChart(Series) composition-root size + slot-style tooltipBody prop, the
LiveChartTransition timed cross-fade, the useMarkers gesture-worklet ref, the
frame/reaction-driven degen + sim effects, the demo formatter bridge, the
pin-once time band, and react-doctor's own self-referential devDependency.

Both projects score 100/100; 652 tests pass.
