+++ doctor.config.ts
@@ -2,33 +2,32 @@ import type { ReactDoctorConfig } from "react-doctor/api";
 
 /**
  * react-native-livechart is a Reanimated + Skia library: data and live values
- * flow through `SharedValue`s and are drawn on the UI thread. Several
- * React-purity lint rules assume an idiomatic React app and fire on patterns
- * that are deliberate (and load-bearing) here.
- *
- * The two rules below are turned off globally because they conflict with the
- * library's core model everywhere it appears. Everything more localized is
- * suppressed per-file in `ignore.overrides` (so the rule keeps protecting the
- * rest of the codebase), each with the reason it's intentional at that spot.
+ * flow through `SharedValue`s and are drawn on the UI thread. The findings
+ * suppressed below are genuine false positives for that architecture (or for
+ * intentional demo code) — each is scoped to the specific files where the
+ * pattern is deliberate, with the reason inline, so every rule keeps protecting
+ * the rest of the codebase. Everything else react-doctor flagged was fixed in
+ * code (see the PR stack): the `.value`→`.get()/.set()` migration, removing
+ * redundant manual memoization, dead-code cleanup, ref/purity/effect fixes,
+ * direct imports, stable keys, FlatList, and lazy ref init.
  */
 const config: ReactDoctorConfig = {
-  rules: {
-    // The library deliberately uses internal barrels (`../hooks`, `../components`)
-    // for organization; public consumers import from the package root barrel.
-    "react-doctor/no-barrel-import": "off",
-
-    // Every flagged list here is a fixed-size render-slot pool (particle slots,
-    // trade-tape labels) or static reference-line config — the array index is the
-    // genuine stable identity, never reorderable user data.
-    "react-doctor/no-array-index-as-key": "off",
-    "react-doctor/no-array-index-key": "off",
-  },
   ignore: {
     overrides: [
       {
+        // Fixed-size render-slot pools (particle slots, trade-tape labels): the
+        // array index IS the stable slot identity, not reorderable user data.
+        files: [
+          "**/components/DegenParticlesOverlay.tsx",
+          "**/components/TradeStreamOverlay.tsx",
+        ],
+        rules: ["react-doctor/no-array-index-as-key"],
+      },
+      {
         // LiveChart / LiveChartSeries are the top-level composition roots; their
-        // size is inherent to a fully-featured chart. The conditional tooltip JSX
-        // prop is fine under React Compiler.
+        // size is inherent to a fully-featured chart component. The optional
+        // `tooltipBody` slot intentionally accepts JSX content as a prop (a
+        // standard slot-style API); React Compiler handles its memoization.
         files: [
           "**/components/LiveChart.tsx",
           "**/components/LiveChartSeries.tsx",
@@ -37,11 +36,6 @@ const config: ReactDoctorConfig = {
           "react-doctor/no-giant-component",
           "react-doctor/jsx-no-jsx-as-prop",
         ],
-      },
-      {
-        // `seriesMetaSig` is a worklet helper intentionally exported for tests.
-        files: ["**/components/SeriesToggleChips.tsx"],
-        rules: ["react-doctor/only-export-components"],
       },
       {
         // Timed cross-fade: mounting/unmounting children on an `active` prop
@@ -53,18 +47,15 @@ const config: ReactDoctorConfig = {
         ],
       },
       {
-        // The gesture is built in useMemo and its worklet runs on the UI thread;
-        // the latest-value ref it reaches is touched on tap, not during render.
+        // The gesture is built with Gesture.Tap() and its worklet runs on the UI
+        // thread; the latest-value ref it reaches is touched on tap, not in render.
         files: ["**/hooks/useMarkers.ts"],
         rules: ["react-hooks-js/refs"],
       },
       {
         // Degen effects are driven by the frame loop / SharedValue signatures, not
         // user events; deps are deliberately narrowed to avoid re-arming the loop.
-        files: [
-          "**/hooks/useDegen.ts",
-          "**/hooks/useMultiSeriesDegen.ts",
-        ],
+        files: ["**/hooks/useDegen.ts", "**/hooks/useMultiSeriesDegen.ts"],
         rules: [
           "react-doctor/exhaustive-deps",
           "react-doctor/no-event-handler",
@@ -72,13 +63,10 @@ const config: ReactDoctorConfig = {
       },
       {
         // Demo trend input: the displayed text mirrors a Reanimated SharedValue
-        // via a reaction; when the formatter changes we re-format the latest
-        // value in an effect. It can't be derived during render (the value
-        // updates asynchronously off the render path), so the derived-state /
-        // pass-data effect rules don't apply here. The Intl formatter and the
-        // `format`/`onValue` callbacks are deliberately memoized (js-hoist-intl
-        // + stable effect/reaction deps), so the redundant-manual-memoization
-        // rule — which conflicts with those — is also waived for this file.
+        // via a reaction; the formatter-change re-format can't be derived during
+        // render (the value updates off the render path). The Intl formatter and
+        // its callbacks are deliberately memoized (js-hoist-intl + stable effect /
+        // reaction deps), which conflicts with the redundant-memoization rule.
         files: ["**/AnimatedTrendTextInput.tsx"],
         rules: [
           "react-doctor/exhaustive-deps",
@@ -88,12 +76,11 @@ const config: ReactDoctorConfig = {
         ],
       },
       {
-        // Demo simulation hook: RNG/bonding state seeded once, history fed through
-        // a callback by design; side effects are reaction-driven, not event-driven.
+        // Demo simulation hook: history is fed through a callback by design and
+        // its side effects are reaction/timer-driven, not user-event-driven.
         files: ["**/useSimulatedChartData.ts"],
         rules: [
           "react-doctor/no-event-handler",
-          "react-doctor/rerender-lazy-ref-init",
           "react-doctor/no-pass-data-to-parent",
         ],
       },
@@ -104,12 +91,7 @@ const config: ReactDoctorConfig = {
         rules: ["react-hooks-js/set-state-in-effect"],
       },
       {
-        // Demo home screen: a short, static list of links — a ScrollView is fine.
-        files: ["app/index.tsx", "**/app/index.tsx"],
-        rules: ["react-doctor/rn-no-scrollview-mapped-list"],
-      },
-      {
-        // react-doctor is a CLI dev tool, never imported — flagging itself.
+        // react-doctor is a CLI dev tool, never imported — it flags itself.
         files: ["package.json", "**/package.json"],
         rules: ["deslop/unused-dev-dependency"],
       },
